backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3
python-dotenv>=1.0.0
pydantic>=2.4.2
treys>=0.1.8
litellm>=0.1.734